Read config before writing
Reading before writing is important because otherwise it would be
possible that the program writes wrong values to the configuration file.

Specific error case:
Create configuration, set date and not project_id, write configuration.
This results in project_id getting a random value that was in the
project_id variable.